Fixed the issues with the drivers, refactoring of the main code for the 2 tivas

+++ Tiva 2/Timer.h
@@ -6,6 +6,8 @@
 
 #define TIMER_LOAD_VAL 0x00F42400 // 16,000,000 in decimal = microcontroller's clock freq. therefore equivalent to one second
 
-void Periodic_Timer_Init(uint8 TimerNo);
+void Periodic_Timer_Init(uint32 TimerNo);
 
 #endif
+
+
